Refactored definitions
