+++ src/definitions/schema.ts
@@ -1,3 +1,10 @@
+/**
+ * Copyright (c) 2017-present, Graphene.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
 import {
   GraphQLObjectType,
   ExecutionResult,
